+++ fe/src/vite-env.d.ts
